requirements.txt
certifi==2018.11.29
chardet==3.0.4
dataclasses==0.6
Django==2.1.4
django-crispy-forms==1.7.2
django-debug-toolbar==1.11
django-filter==2.0.0
djangorestframework==3.9.0
djongo==1.2.31
future==0.17.1
idna==2.8
iso8601==0.1.12
Markdown==3.0.1
Pillow==5.4.1
pymongo==3.7.2
pyserial==3.4
pytz==2018.7
pywin32==224
PyYAML==3.13
requests==2.21.0
six==1.12.0
sqlparse==0.2.4
urllib3==1.24.1
